添加 删除消息逻辑判断 修正文字错误 添加keep_awake_url变量

+++ tgfilestream/telegram.py
@@ -14,8 +14,11 @@
 # You should have received a copy of the GNU Affero General Public License
 # along with this program.  If not, see <https://www.gnu.org/licenses/>.'
 import logging
-
+from typing import cast
+import asyncio
 from telethon import TelegramClient, events
+from telethon.tl.custom import Message
+from telethon.tl.types import InputPeerChannel, InputPeerChat, InputPeerUser
 
 from .config import link_prefix, api_id, api_hash, allowed_user, max_file_size, admin_id, session
 from .paralleltransfer import ParallelTransferrer
@@ -28,21 +31,52 @@ client = TelegramClient(session, api_id, api_hash)
 transfer = ParallelTransferrer(client)
 
 
-@client.on(events.NewMessage)
+def new_message_filter(message_text):
+    return not str(message_text).startswith('/start')
+
+
+@client.on(events.NewMessage(pattern='/start'))
+async def handle_start(evt: events.NewMessage.Event) -> None:
+    c = await evt.reply('send me an image to host it on web')
+    # log.debug(c)
+    await asyncio.sleep(5)
+    await client.delete_messages(evt.input_chat, [c.id])
+    await evt.delete()
+    raise events.StopPropagation
+
+
+@client.on(events.NewMessage(pattern=new_message_filter))
 async def handle_message(evt: events.NewMessage.Event) -> None:
-    if str(evt.from_id) not in allowed_user or str(evt.chat_id) not in allowed_user:
+    if '*' in allowed_user and not evt.is_private:
+        return
+    if not (str(evt.from_id) in allowed_user and str(evt.chat_id) in allowed_user) and '*' not in allowed_user:
         log.info(f'user {evt.from_id} or {evt.chat_id} not allowed to use this bot')
-        await evt.delete()
+        if evt.is_private:
+            await evt.delete()
         return
     if str(evt.message.message).startswith('/del') and evt.reply_to_msg_id is not None:
-        log.debug(evt)
-        await client.delete_messages(evt.input_chat, [evt.reply_to_msg_id])
+        if bool(int(evt.is_channel)) and bool(int(evt.is_group)):
+            peer = InputPeerChat(chat_id=int(evt.chat_id))
+        else:
+            if bool(int(evt.is_group)):
+                peer = InputPeerChat(chat_id=int(evt.chat_id))
+            elif bool(int(evt.is_channel)):
+                peer = InputPeerChannel(channel_id=int(evt.chat_id), access_hash=0)
+            else:
+                peer = InputPeerUser(user_id=int(evt.chat_id), access_hash=0)
+        c = cast(Message, await client.get_messages(entity=peer, ids=evt.reply_to_msg_id))
+        me = await client.get_me()
+        reply_msg = cast(Message, await c.get_reply_message())
+        log.debug(f'msg_from={c.from_id},evt_from={evt.from_id},c is reply={c.is_reply}'
+                  f',reply_msg_from={reply_msg.from_id if reply_msg is not None else 0}')
+        if c.from_id == evt.from_id or (c.from_id == me.id and c.is_reply):
+            if (reply_msg is not None and reply_msg.from_id == evt.from_id) or reply_msg is None:
+                await client.delete_messages(evt.input_chat, [evt.reply_to_msg_id])
         await evt.delete()
     else:
         if not evt.file:
             log.info('not evt.file')
-            if not str(evt.message.message).startswith('/start'):
-                await evt.delete()
+            await evt.delete()
             return
         try:
             ret = get_media_meta(evt.media)
